requirements.txt
pyairtable
requests
pandas
python-dotenv
databricks-sdk
pytest
black
ruff
